Added Delete button
Delete button added in each book in the row, to remove the specific book from the list.
The delete button, created when a new book row is created,
is inserted in a tabular data cell, which is part of the tabular row of the book.
The Button click event is bind with a function to remove the parent parent node (the entire row)